+++ CobaltTests/CobaltTests.swift
@@ -7,29 +7,32 @@
 //
 
 import XCTest
-import Nimble
-import Promises
+import Logging
 import Alamofire
+import RxSwift
+import RxCocoa
 @testable import Cobalt
 
-
 class CobaltTests: XCTestCase {
+    lazy var disposeBag = DisposeBag()
     lazy var config = Config {
-        $0.clientID = "id"
-        $0.clientSecret = "secret"
-        $0.clientAuthorization = .requestBody
+        $0.authentication.clientID = "id"
+        $0.authentication.clientSecret = "secret"
+        $0.authentication.authorization = .requestBody
+        $0.logging.logger = Logger(label: "com.esites.cobalt-test")
+        $0.logging.logger?.logLevel = .trace
+        
         $0.host = "https://reqres.in"
-        $0.logger = TestLogger()
     }
     lazy var client = Cobalt.Client(config: self.config)
-
-    override func setUp() {
-        super.setUp()
-        Nimble.AsyncDefaults.Timeout = 15
-        Nimble.AsyncDefaults.PollInterval = 0.1
-    }
     
     override func tearDown() {
         super.tearDown()
+    }
+    
+    func waitUntil(_ handler: @escaping (((() -> Void)?) -> Void)) {
+        let expectation = self.expectation(description: "test")
+        handler({ expectation.fulfill() })
+        waitForExpectations(timeout: 15, handler: nil)
     }
 }