+++ Package.swift
@@ -4,18 +4,18 @@ import PackageDescription
 let package = Package(
     name: "Cobalt",
     platforms: [
-        .iOS(.v9),
+        .iOS(.v10),
     ],
     products: [
         .library(name: "Cobalt", targets: ["Cobalt"]),
         .library(name: "CobaltCache", targets: ["CobaltCache"])
     ],
     dependencies: [
-        .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "4.0.0")),
+        .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.2.1")),
         .package(url: "https://github.com/SwiftyJSON/SwiftyJSON.git", .upToNextMajor(from: "5.0.0")),
-        .package(url: "https://github.com/ReactiveX/RxSwift.git", .upToNextMajor(from: "5.0.0")),
-        .package(url: "https://github.com/google/promises.git", .upToNextMajor(from: "1.0.0")),
-        .package(url: "https://github.com/kishikawakatsumi/KeychainAccess.git", .upToNextMajor(from: "3.0.0"))
+        .package(url: "https://github.com/ReactiveX/RxSwift.git", .upToNextMajor(from: "6.0.0")),
+        .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
+        .package(url: "https://github.com/kishikawakatsumi/KeychainAccess.git", .upToNextMajor(from: "4.1.0"))
     ],
     targets: [
         .target(
@@ -24,8 +24,9 @@ let package = Package(
                 "Alamofire",
                 "SwiftyJSON",
                 "RxSwift",
-                "Promises",
-                "KeychainAccess"
+                "RxCocoa",
+                "KeychainAccess",
+                "Logging"
             ],
             path: "Sources/Core"
         ),