+++ Sources/Core/Base/Client.swift
@@ -8,9 +8,10 @@
 
 import Foundation
 import RxSwift
-import Promises
+import RxCocoa
 import SwiftyJSON
 import Alamofire
+import Logging
 
 open class Client: ReactiveCompatible {
 
@@ -25,7 +26,17 @@ open class Client: ReactiveCompatible {
 
     fileprivate lazy var queue = RequestQueue(client: self)
 
-    var authorizationGrantTypeSubject = BehaviorSubject<OAuthenticationGrantType?>(value: nil)
+    lazy var authorizationGrantTypeSubject: BehaviorSubject<OAuthenticationGrantType?> = {
+        var value: OAuthenticationGrantType?
+        if let grantTypeRawValue = UserDefaults.standard.string(forKey: "OAuthenticationGrantType"),
+            let grantType = OAuthenticationGrantType(rawValue: grantTypeRawValue) {
+            value = grantType
+        }
+        
+        return BehaviorSubject<OAuthenticationGrantType?>(value: value)
+    }()
+
+    private let disposeBag = DisposeBag()
 
     var authorizationGrantType: OAuthenticationGrantType? {
         do {
@@ -36,14 +47,18 @@ open class Client: ReactiveCompatible {
     }
 
     public var accessToken: AccessToken? {
-        guard let host = config.host else {
+        guard let host = authenticationHost else {
             return nil
         }
         return AccessToken(host: host)
     }
 
     var logger: Logger? {
-        return config.logger
+        return config.logging.logger
+    }
+    
+    var authenticationHost: String? {
+        return config.authentication.host ?? config.host
     }
 
     @objc
@@ -55,6 +70,11 @@ open class Client: ReactiveCompatible {
     required public init(config: Config) {
         self.config = config
         service.logger = logger
+        
+        authorizationGrantTypeSubject.subscribe(onNext: { grantType in
+            UserDefaults.standard.set(grantType?.rawValue, forKey: "OAuthenticationGrantType")
+            _ = UserDefaults.standard.synchronize()
+        }).disposed(by: disposeBag)
     }
 
     // MARK: - Request functions
@@ -65,12 +85,12 @@ open class Client: ReactiveCompatible {
     /// - Parameters:
     ///   - `request`: The `Request` object
     ///   - `handler`: The closure to call when the request is finished
-    open func request(_ request: Request, handler: @escaping ((Alamofire.Result<JSON>) -> Void)) {
-        self.request(request).then { json in
+    open func request(_ request: Request, handler: @escaping ((Result<JSON, Swift.Error>) -> Void)) {
+        self.request(request).subscribe(onSuccess: { json in
             handler(.success(json))
-        }.catch { error in
+        }, onFailure: { error in
             handler(.failure(error))
-        }
+        }).disposed(by: disposeBag)
     }
 
     /// Make a promise requst
@@ -79,10 +99,10 @@ open class Client: ReactiveCompatible {
     ///   - `request`: The `Request` object
     ///
     /// - Returns: `Promise<JSON>`
-    open func request(_ request: Request) -> Promise<JSON> {
+    open func request(_ request: Request) -> Single<JSON> {
         // Strip slashes to form a valid urlString
-        guard var host = (request.host ?? config.host) else {
-            return Promise(Error.invalidRequest("Missing 'host'"))
+        guard let host = (request.host ?? config.host) else {
+            return Single<JSON>.error(Error.invalidRequest("Missing 'host'").set(request: request))
         }
 
         // If the client is authenticating with OAuth.
@@ -90,17 +110,10 @@ open class Client: ReactiveCompatible {
         // So we add it to the `RequestQueue`
         if authProvider.isAuthenticating && request.requiresOAuthentication {
             queue.add(request)
-            return queue.promise(of: request) ?? Promise(Error.unknown())
+            return queue.single(of: request) ?? Single<JSON>.error(Error.unknown().set(request: request))
         }
         
-        if host.hasSuffix("/") {
-            host = String(host.dropLast())
-        }
-        var path = request.path
-        if path.hasPrefix("/") {
-            path = String(path.dropFirst())
-        }
-        let urlString = host + "/" + path
+        let urlString = String.combined(host: host, path: request.path)
 
         // Define encoding
         let encoding: ParameterEncoding
@@ -116,25 +129,29 @@ open class Client: ReactiveCompatible {
         request.urlString = urlString
         
         // 1. We (optionally) (pre-)authorize the request
-        return firstly {
+        return Single<Void>.just(()).flatMap { [authProvider] in 
             return try authProvider.authorize(request: request)
 
         // 2. We actually send the request with Alamofire
-        }.then { newRequest in
+        }.flatMap { [weak self] newRequest in
+            guard let self = self else {
+                return Single<JSON>.never()
+            }
             return self._request(newRequest)
 
         // 3. If for some reason an error occurs, we check with the auth-provider if we need to retry
-        }.recover { error -> Promise<JSON> in
-            self.queue.removeFirst()
-            return try self.authProvider.recover(from: error, request: request)
+        }.catch { [weak self, authProvider] error -> Single<JSON> in
+            self?.queue.removeFirst()
+            return try authProvider.recover(from: error, request: request)
 
         // 4. If any other requests are queued, fire up the next one
-        }.always {
+        }.flatMap { [queue] json -> Single<JSON> in
             // When a request is finished, no matter if its succesful or not
             // We try to clear th queue
             if request.requiresOAuthentication {
-                self.queue.next()
+                queue.next()
             }
+            return Single.just(json)
         }
     }
 
@@ -146,39 +163,55 @@ open class Client: ReactiveCompatible {
 
     }
 
-    private func _request(_ request: Request) -> Promise<JSON> {
+    private func _request(_ request: Request) -> Single<JSON> {
         let requestID = self.requestID
         self.requestID += 1
         if self.requestID == 100 {
             self.requestID = 1
         }
         var loggingOptions: [String: KeyLoggingOption] = [:]
-        if config.maskTokens {
+        if config.logging.maskTokens {
             loggingOptions["Authorization"] = .halfMasked
         }
 
-        if !request.useHeaders.keys.isEmpty {
-            let headersDictionary = dictionaryForLogging(request.useHeaders, options: loggingOptions)
-            logger?.verbose("#\(requestID) Headers: \(headersDictionary ?? [:])")
+        let ignoreLoggingRequest: Bool
+        if let logReq = request.loggingOption?.request?["*"], case KeyLoggingOption.ignore = logReq {
+            ignoreLoggingRequest = true
+        } else {
+            ignoreLoggingRequest = false
         }
 
-        let loggingParameters = dictionaryForLogging(request.parameters,
-                                                     options: request.loggingOption?.request)
+        let ignoreLoggingResponse: Bool
+        if let logRes = request.loggingOption?.response?["*"], case KeyLoggingOption.ignore = logRes {
+            ignoreLoggingResponse = true
+        } else {
+            ignoreLoggingResponse = false
+        }
 
-        logger?.request("#\(requestID) " + request.httpMethod.rawValue,
-                        request.urlString,
-                        loggingParameters?.flatJSONString ?? "")
+        if !request.useHeaders.isEmpty, !ignoreLoggingRequest {
+            let headersDictionary = Helpers.dictionaryForLogging(request.useHeaders.dictionary, options: loggingOptions)
+            logger?.notice("#\(requestID) Headers: \(headersDictionary ?? [:])")
+        }
+
+        if !ignoreLoggingRequest {
+            let loggingParameters = Helpers.dictionaryForLogging(request.parameters,
+                                                         options: request.loggingOption?.request)
+
+            logger?.trace("[REQ] #\(requestID) \(request.httpMethod.rawValue) \(request.urlString) \(loggingParameters?.flatJSONString ?? "")",  metadata: [ "tag": "api" ])
+        }
         startRequest(request)
 
         service.currentRequest = request
 
         // Check to see if the cache engine should handle it
         if !service.shouldPerformRequestAfterCacheCheck(), let json = service.json {
-            _responseParsing(json: json, request: request, requestID: requestID)
-            return Promise(json)
+            if !ignoreLoggingResponse {
+                _responseParsing(json: json, request: request, requestID: requestID)
+            }
+            return Single<JSON>.just(json)
         }
 
-        return Promise<JSON>(on: .main) { [weak self] fulfill, reject in
+        return Single<JSON>.create { [weak self] observer in
             let dataRequest: DataRequest
             if let data = request.body {
                 var urlRequest = URLRequest(url: URL(string: request.urlString)!)
@@ -197,54 +230,74 @@ open class Client: ReactiveCompatible {
             }
             
             dataRequest.validate()
-
-                .validate()
                 .responseJSON { response in
-
                     let statusCode = response.response?.statusCode ?? 500
                     self?.finishRequest(request, response: response.response)
                     let statusString = HTTPURLResponse.localizedString(forStatusCode: statusCode)
-                    self?.logger?.verbose("#\(requestID) HTTP Status: \(statusCode) ('\(statusString)')")
+                    if !ignoreLoggingResponse {
+                        self?.logger?.notice("#\(requestID) HTTP Status: \(statusCode) ('\(statusString)')")
+                    }
+
                     var json: JSON?
                     if let data = response.data {
                         json = JSON(data)
                         self?.service.json = json
                         self?.service.optionallyWriteToCache()
-                        self?._responseParsing(json: json, request: request, requestID: requestID)
+                        if !ignoreLoggingResponse {
+                            self?._responseParsing(json: json, request: request, requestID: requestID)
+                        }
                     }
 
                     if let error = response.error {
-                        self?.logger?.error("#\(requestID) Original: \(error)")
-                        let apiError = Error(from: error, json: json)
-                        self?.logger?.error("#\(requestID) Error: \(apiError)")
-                        reject(apiError)
+                        let apiError = Error(from: error, json: json).set(request: request)
+                        if !ignoreLoggingResponse {
+                            self?.logger?.error("#\(requestID) Original: \(error)")
+                            self?.logger?.error("#\(requestID) Error: \(apiError)")
+                        }
+                        observer(.failure(apiError))
                         return
                     }
 
                     guard let responseJSON = json else {
-                        reject(Error.empty)
+                        observer(.failure(Error.empty.set(request: request)))
                         return
                     }
-
-                    fulfill(responseJSON)
+                    observer(.success(responseJSON))
             }
+            return Disposables.create()
         }
     }
 
     private func _responseParsing(json: JSON?, request: Request, requestID: Int) {
-        let dictionary = dictionaryForLogging(json?.dictionaryObject ?? [:], options: request.loggingOption?.response)
-        logger?.response("#\(requestID) " + request.httpMethod.rawValue, request.urlString, dictionary?.flatJSONString ?? "")
+        var responseString: String?
+        if let dictionaryObject = json?.dictionaryObject {
+            let dictionary = Helpers.dictionaryForLogging(dictionaryObject, options: request.loggingOption?.response)
+            responseString = dictionary?.flatJSONString
+            
+        } else {
+            responseString = json?.flatString
+        }
+        
+        logger?.trace("[RES] #\(requestID) \(request.httpMethod.rawValue)  \(request.urlString) \(responseString ?? "")", metadata: [ "tag": "api"])
     }
 
     // MARK: - Login
     // --------------------------------------------------------
 
-    public func login(username: String, password: String) -> Promise<Void> {
+    open func login(username: String, password: String) -> Single<Void> {
         let parameters = [
             "username": username,
             "password": password
         ]
         return authProvider.sendOAuthRequest(grantType: .password, parameters: parameters)
+    }
+    
+    open func startAuthorizationFlow(scope: [String], redirectUri: String) -> Single<AuthorizationCodeRequest> {
+        return authProvider.createAuthorizationCodeRequest(scope: scope, redirectUri: redirectUri)
+    }
+    
+    open func requestTokenFromAuthorizationCode(initialRequest request: AuthorizationCodeRequest, code: String) -> Single<Void> {
+        return authProvider.requestTokenFromAuthorizationCode(initialRequest: request, code: code)
     }
     
     /// Handle the result of a manual login call
@@ -261,7 +314,7 @@ open class Client: ReactiveCompatible {
                          accessToken: String,
                          refreshToken: String,
                          expireDate: Date) {
-        guard let host = config.host else {
+        guard let host = authenticationHost else {
             fatalError("No valid host set in the config")
         }
         
@@ -270,7 +323,7 @@ open class Client: ReactiveCompatible {
 
     public func clearAccessToken(forHost host: String? = nil) {
         authorizationGrantTypeSubject.onNext(nil)
-        guard let host = (host ?? config.host) else {
+        guard let host = (host ?? authenticationHost) else {
             fatalError("No host given, nor a valid host set in the Cobalt.Config")
         }
         AccessToken(host: host).clear()
@@ -279,74 +332,6 @@ open class Client: ReactiveCompatible {
 
 // MARK: - Helpers
 // --------------------------------------------------------
-
-extension Client {
-    func dictionaryForLogging(_ parameters: [String: Any]?,
-                              options: [String: KeyLoggingOption]?) -> [String: Any]? {
-        guard let theParameters = parameters, let options = options else {
-            return parameters
-        }
-        return _mask(parameters: theParameters, options: options)
-    }
-
-    fileprivate func _mask(parameters: [String: Any],
-                           options: [String: KeyLoggingOption],
-                           path: String = "") -> [String: Any] {
-        var logParameters: [String: Any] = [:]
-        for (key, value) in parameters {
-            let type = options["\(path)\(key)"] ?? .default
-            if let dictionary = value as? [String: Any] {
-                logParameters[key] = _mask(parameters: dictionary, options: options, path: "\(path)\(key).")
-                continue
-            }
-            guard let string = Client.mask(string: value, type: type) else {
-                continue
-            }
-            logParameters[key] = string
-        }
-        return logParameters
-    }
-
-    class func mask(string value: Any?, type: KeyLoggingOption) -> Any? {
-        guard let value = value else {
-            return nil
-        }
-        switch type {
-        case .halfMasked:
-            guard let stringValue = value as? String, !stringValue.isEmpty else {
-                return value
-            }
-            let length = Int(floor(Double(stringValue.count) / 2.0))
-            let startIndex = stringValue.startIndex
-            let midIndex = stringValue.index(startIndex, offsetBy: length)
-            return String(describing: stringValue[startIndex..<midIndex]) + "***"
-
-        case .ignore:
-            return nil
-
-        case .replaced(let string):
-            return string
-
-        case .masked:
-            return "***"
-
-        case .shortened:
-            guard let stringValue = value as? String else {
-                fallthrough
-            }
-            if stringValue.count > 128 {
-                let startIndex = stringValue.startIndex
-                let endIndex = stringValue.index(startIndex, offsetBy: 128)
-                return String(describing: stringValue[startIndex..<endIndex]) + "..."
-            } else {
-                return value
-            }
-            
-        default:
-            return value
-        }
-    }
-}
 
 extension Reactive where Base: Client {
     public var authorizationGrantType: Observable<OAuthenticationGrantType?> {